chore(security): moved instances of insertion to server side and moved around function locations
