+++ src/store/uploadedFile.ts
@@ -1,4 +1,4 @@
-import { UploadedFile } from "@/components/custom-uploader";
+import { UploadedFile } from "@/types/file";
 import { create } from "zustand";
 
 type UploadedFileStore = {
@@ -6,6 +6,7 @@ type UploadedFileStore = {
   addUploadedFile: (file: UploadedFile) => void;
   removeUploadedFile: (fileKey: string) => void;
   isAtFileLimit: () => boolean;
+  clearUploadedFiles: () => void;
 };
 
 export const useUploadedFileStore = create<UploadedFileStore>((set, get) => ({
@@ -17,6 +18,7 @@ export const useUploadedFileStore = create<UploadedFileStore>((set, get) => ({
       uploadedFiles: state.uploadedFiles.filter((file) => file.key !== key),
     })),
   isAtFileLimit: () => get().uploadedFiles.length >= MAX_FILES,
+  clearUploadedFiles: () => set({ uploadedFiles: [] }),
 }));
 
 export const MAX_FILES = 3;
